[refactor] Move command from action folder
