--- src/actions/command/changelog.ts
@@ -1,10 +1,0 @@
-/**
- * @external
- * @module commander.command
- */
-
-import winston from "winston";
-
-export default (..._: any[]) => {
-  winston.info("show changelogs");
-};
