+++ src/command/config.ts
@@ -4,9 +4,9 @@
  */
 
 import { log } from "winston";
-import { ThrowIf } from "../../helpers/action";
-import Config from "../../models/Config";
-import { WrapTM, WrapTMC } from "../../models/LoggerWrapper";
+import { ThrowIf } from "../helpers/action";
+import Config from "../models/Config";
+import { WrapTM, WrapTMC } from "../models/LoggerWrapper";
 
 /**
  * This is configuration command, This command able to show the config path
